Database 클래스 별도 생성 및 적용
